fixing output of files
Fixing code to create/write the assignment files for name draws in a given year.  There was a bug if the 'gifter' did not have an address, order of names in output returned was switched.  Also introduced function to parse the address text stored in the database to something more easily readable by a human.